+++ nlp-pipelines-conceptmapper/src/test/java/edu/ucdenver/ccp/nlp/pipelines/conceptmapper/dictmod/PRDictionaryEntryModifierTest.java
@@ -1,4 +1,4 @@
-package edu.ucdenver.ccp.nlp.pipelines.evaluation.craft.conceptmapper;
+package edu.ucdenver.ccp.nlp.pipelines.conceptmapper.dictmod;
 
 /*
  * #%L
